requirements.txt
google-api-python-client==2.70.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
python-dateutil==2.8.2
pytest==7.4.2